// frontend/src/api/generated/v1/models/clientOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wedding Management API (Ninja)
 * OpenAPI spec version: 1.0.0
 */

/**
 * Schema de saída para representação de cliente.
 */
export interface ClientOut {
  uuid: string;
  name: string;
  cpf?: string;
  email?: string;
  phone?: string;
  notes?: string;
  created_at: string;
  updated_at: string;
}